+++ CHANGELOG.md
@@ -11,6 +11,8 @@ Enhancements:
 
 * [#338](https://github.com/BurntSushi/jiff/pull/338):
 Add support for the `%c`, `%r`, `%X` and `%x` conversion specifiers.
+* [#341](https://github.com/BurntSushi/jiff/issues/341):
+Add support for `%q` in `jiff::fmt::strtime` (prints quarter of year).
 * [#344](https://github.com/BurntSushi/jiff/issues/344):
 Add support for `%N` in `jiff::fmt::strtime` (alias for `%9f`).